+++ sofa/src/components/Layout/main-layout.css
@@ -1,4 +1,4 @@
-.userpage {
+.homepage {
     position: relative;
     display: flex;
     flex-direction: column; /* 레이아웃 조정 */
@@ -25,7 +25,7 @@
 }
 
 /* 사이드 메뉴가 열릴 때 */
-.userpage.menu-open {
+.homepage.menu-open {
   width: calc(100% - 16.5rem); /* 사이드 메뉴의 너비만큼 화면이 줄어듬 */
   margin-left: 15.5rem;  /* 사이드 메뉴의 너비만큼 왼쪽 여백을 추가 */
 }
