+++ sofa/src/pages/onBoarding-layout.css
@@ -3,7 +3,8 @@
     white-space: nowrap;    /* 화면 크기에 따른 자동 줄바꿈 방지 */
 }
 
-.main-style{    /* 컴포넌트 가운데 정렬 */
+ /* 컴포넌트 가운데 정렬 */
+.main-style{
     user-select: none;
     display:flex;
     flex-direction: column;
